requirements.txt
Django==4.2.25
djangorestframework
drf-yasg
opencv-python
numpy
Pillow
torch
torchvision
scipy
python-dotenv